refactor(:recycle:) - post 삭제 시, async로 event 처리
