+++ src/main/java/kr/ac/hs/selab/board/domain/event/BoardEvent.java
@@ -1,5 +1,6 @@
-package kr.ac.hs.selab.board.domain;
+package kr.ac.hs.selab.board.domain.event;
 
+import kr.ac.hs.selab.board.domain.Board;
 import lombok.Value;
 
 @Value(staticConstructor = "of")
