+++ src/main/java/kr/ac/hs/selab/board/eventListener/BoardEventListener.java
@@ -1,6 +1,6 @@
-package kr.ac.hs.selab.board.event;
+package kr.ac.hs.selab.board.eventListener;
 
-import kr.ac.hs.selab.board.domain.BoardEvent;
+import kr.ac.hs.selab.board.domain.event.BoardEvent;
 import kr.ac.hs.selab.comment.application.CommentService;
 import kr.ac.hs.selab.post.application.PostService;
 import lombok.RequiredArgsConstructor;
@@ -10,8 +10,8 @@ import org.springframework.transaction.event.TransactionalEventListener;
 
 import javax.transaction.Transactional;
 
-@Component
 @RequiredArgsConstructor
+@Component
 public class BoardEventListener {
     private final PostService postService;
     private final CommentService commentService;
@@ -19,7 +19,7 @@ public class BoardEventListener {
     @Async
     @Transactional(value = Transactional.TxType.REQUIRES_NEW)
     @TransactionalEventListener
-    public void removeBoard(BoardEvent boardEvent) {
+    public void deleteByBoard(BoardEvent boardEvent) {
         var board = boardEvent.getBoard();
         commentService.deleteByPosts(
                 postService.findPostsByBoard(board)
