+++ src/main/java/kr/ac/hs/selab/post/domain/event/PostEvent.java
@@ -1,0 +1,9 @@
+package kr.ac.hs.selab.post.domain.event;
+
+import kr.ac.hs.selab.post.domain.Post;
+import lombok.Value;
+
+@Value(staticConstructor = "of")
+public class PostEvent {
+    Post post;
+}
