+++ src/main/java/kr/ac/hs/selab/post/facade/PostFacade.java
@@ -2,12 +2,11 @@ package kr.ac.hs.selab.post.facade;
 
 import kr.ac.hs.selab.board.application.BoardService;
 import kr.ac.hs.selab.board.domain.Board;
-import kr.ac.hs.selab.board.domain.BoardEvent;
-import kr.ac.hs.selab.comment.application.CommentService;
 import kr.ac.hs.selab.member.application.MemberService;
 import kr.ac.hs.selab.member.domain.Member;
 import kr.ac.hs.selab.post.application.PostService;
 import kr.ac.hs.selab.post.domain.Post;
+import kr.ac.hs.selab.post.domain.event.PostEvent;
 import kr.ac.hs.selab.post.dto.PostCreateDto;
 import kr.ac.hs.selab.post.dto.response.PostFindByBoardResponse;
 import kr.ac.hs.selab.post.dto.response.PostResponse;
@@ -17,13 +16,13 @@ import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
 @RequiredArgsConstructor
+@Transactional(readOnly = true)
 @Component
 public class PostFacade {
     private final ApplicationEventPublisher publisher;
     private final MemberService memberService;
     private final BoardService boardService;
     private final PostService postService;
-    private final CommentService commentService;
 
     @Transactional
     public PostResponse create(PostCreateDto postDto) {
@@ -38,9 +37,11 @@ public class PostFacade {
         return postService.findPostsResponseByBoard(board);
     }
 
+    @Transactional
     public PostResponse delete(Long postId) {
         Post post = postService.delete(postId);
-        commentService.deleteByPost(post);
+        publisher.publishEvent(PostEvent.of(post));
+
         return new PostResponse(post.getId());
     }
 }
